Read boot reason (for offline charging) to skip recovery boot

+++ combinedroot/init.sh
@@ -29,46 +29,51 @@ busybox mount -t proc proc /proc
 busybox mount -t sysfs sysfs /sys
 busybox mount -t ext4 ${BOOTREC_CACHE} /cache
 
-# trigger ON amber LED
-busybox echo ${BOOTREC_RED_LED_ON} > ${BOOTREC_CONTROL_LED}
-busybox echo ${BOOTREC_GREEN_LED_OFF} > ${BOOTREC_CONTROL_LED}
-busybox echo ${BOOTREC_BLUE_LED_ON} > ${BOOTREC_CONTROL_LED}
-busybox echo ${BOOTREC_BUTTON_LED_ON} > ${BOOTREC_CONTROL_LED}
-
-# keycheck
-busybox cat ${BOOTREC_EVENT} > /dev/keycheck&
-busybox sleep 3
-
-# trigger OFF amber LED
-busybox echo ${BOOTREC_RED_LED_OFF} > ${BOOTREC_CONTROL_LED}
-busybox echo ${BOOTREC_GREEN_LED_OFF} > ${BOOTREC_CONTROL_LED}
-busybox echo ${BOOTREC_BLUE_LED_OFF} > ${BOOTREC_CONTROL_LED}
-busybox echo ${BOOTREC_BUTTON_LED_OFF} > ${BOOTREC_CONTROL_LED}
-
 # android ramdisk
 load_image=/sbin/ramdisk.cpio
 
-# boot decision
-if [ -s /dev/keycheck -o -e /cache/recovery/boot ]
+busybox grep -Eq "^64|128$" /proc/sys/kernel/boot_reason
+
+if [[ $? != 0 ]]
 then
-	busybox echo 'RECOVERY BOOT' >>boot.txt
-	busybox rm -fr /cache/recovery/boot
-	# trigger blue led
-	busybox echo ${BOOTREC_RED_LED_OFF} > ${BOOTREC_CONTROL_LED}
+	# trigger ON amber LED
+	busybox echo ${BOOTREC_RED_LED_ON} > ${BOOTREC_CONTROL_LED}
 	busybox echo ${BOOTREC_GREEN_LED_OFF} > ${BOOTREC_CONTROL_LED}
 	busybox echo ${BOOTREC_BLUE_LED_ON} > ${BOOTREC_CONTROL_LED}
-	# recovery ramdisk
-	load_image=/sbin/ramdisk-recovery.cpio
-else
-	busybox echo 'ANDROID BOOT' >>boot.txt
-	# poweroff LED
+	busybox echo ${BOOTREC_BUTTON_LED_ON} > ${BOOTREC_CONTROL_LED}
+
+	# keycheck
+	busybox cat ${BOOTREC_EVENT} > /dev/keycheck&
+	busybox sleep 3
+
+	# trigger OFF amber LED
 	busybox echo ${BOOTREC_RED_LED_OFF} > ${BOOTREC_CONTROL_LED}
 	busybox echo ${BOOTREC_GREEN_LED_OFF} > ${BOOTREC_CONTROL_LED}
 	busybox echo ${BOOTREC_BLUE_LED_OFF} > ${BOOTREC_CONTROL_LED}
-fi
+	busybox echo ${BOOTREC_BUTTON_LED_OFF} > ${BOOTREC_CONTROL_LED}
 
-# kill the keycheck process
-busybox pkill -f "busybox cat ${BOOTREC_EVENT}"
+	# boot decision
+	if [ -s /dev/keycheck -o -e /cache/recovery/boot ]
+	then
+		busybox echo 'RECOVERY BOOT' >>boot.txt
+		busybox rm -fr /cache/recovery/boot
+		# trigger blue led
+		busybox echo ${BOOTREC_RED_LED_OFF} > ${BOOTREC_CONTROL_LED}
+		busybox echo ${BOOTREC_GREEN_LED_OFF} > ${BOOTREC_CONTROL_LED}
+		busybox echo ${BOOTREC_BLUE_LED_ON} > ${BOOTREC_CONTROL_LED}
+		# recovery ramdisk
+		load_image=/sbin/ramdisk-recovery.cpio
+	else
+		busybox echo 'ANDROID BOOT' >>boot.txt
+		# poweroff LED
+		busybox echo ${BOOTREC_RED_LED_OFF} > ${BOOTREC_CONTROL_LED}
+		busybox echo ${BOOTREC_GREEN_LED_OFF} > ${BOOTREC_CONTROL_LED}
+		busybox echo ${BOOTREC_BLUE_LED_OFF} > ${BOOTREC_CONTROL_LED}
+	fi
+
+	# kill the keycheck process
+	busybox pkill -f "busybox cat ${BOOTREC_EVENT}"
+fi
 
 # unpack the ramdisk image
 busybox cpio -i < ${load_image}
